edit category and main page

+++ app/views/main.php
@@ -2,8 +2,6 @@
 session_start();
 include "app/views/header.php"; 
 ?>
-
-   <!-- end Navbar -->
    <?php
    echo "<div class='col-md-12'>";
    if($action=='added'){
@@ -42,8 +40,8 @@ echo "</div>";
    }
 echo "</div>";
 ?>
- <!-- Start  silder -->
- <div class="slider">
+   <!-- Start  silder -->
+<div class="slider">
     <div id="mysild" class="carousel slide " data-ride="carousel" >
 
          <div class="carousel-inner">
@@ -76,6 +74,23 @@ echo "</div>";
  </div>
 
 
+
+
+ <!-- End  silder -->
+
+<!-- Product slider Start -->
+<!--div class="slider">
+  <div class="container">
+    <div class="owl-carousel">
+      <div class="item"><img src="app/assets/images/Acer.jpg"></div>
+      <div class="item"><img src="app/assets/images/lenovo4.png"></div>
+      <div class="item"><img src="app/assets/images/Acer_Chromebook_315_01_32ff69f880-300x274.jpg"></div>
+      <div class="item"><img src="app/assets/images/ASUS-Chromebook-C423-1-300x202 (1).jpg"></div>
+      <div class="item"><img src="app/assets/images/htc-desire-12_990b.jpg"></div>
+    </div>
+     
+  </div>
+</div-->
 
 <div id="demo" class="carousel slide" data-ride="carousel">
 
@@ -118,17 +133,24 @@ echo "</div>";
                  
               </div>
               <div class="product-content">
-                  <h3 class="title"><a href="#"> <?php  echo $row->product_name ?></a></h3>
+                  <h3 class="title"><a href="#"> <?php  echo $row->product_short_desc ?></a></h3>
                   <div class="price">
                   $<?= $row->product_price ?>
                     <!-- <?php  echo $row->product_price ?> -->
                   </div>
                   <?PHP
-                    
+                    // add to cart button
+         // product id for javascript access
          echo "<div class='product-id display-none'></div>";
-        
+        if(array_key_exists($id, $_SESSION['cart'])){
+            // echo "<a href='main/displayShopingCartItems' class='btn btn-success w-100-pct'>";
+            //     echo "Update Cart";
             echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
-       
+
+            echo "</a>";
+        }else{
+            echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
+        }
     ?>
                 </div>
           </div>
@@ -167,7 +189,7 @@ echo "</div>";
                  
               </div>
               <div class="product-content">
-                  <h3 class="title"><a href="#"> <?php  echo $row->product_name ?></a></h3>
+                  <h3 class="title"><a href="#"> <?php  echo $row->product_short_desc ?></a></h3>
                   <div class="price">
                   
                     <?php  echo $row->product_price ?>
@@ -176,8 +198,15 @@ echo "</div>";
                     // add to cart button
          // product id for javascript access
          echo "<div class='product-id display-none'></div>";
-        echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
-        
+        if(array_key_exists($id, $_SESSION['cart'])){
+            // echo "<a href='main/displayShopingCartItems' class='btn btn-success w-100-pct'>";
+            //     echo "Update Cart";
+            echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
+
+            echo "</a>";
+        }else{
+            echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
+        }
     ?>
                 </div>
           </div>
@@ -212,12 +241,12 @@ echo "</div>";
                   <ul class="social">
                       <li><a href="main/product_details?action=product_details&product_id=<?PHP echo $id?>" data-tip="Quick View"><i class="fa fa-eye"></i></a></li>
                       <li><a href="main/wishlist" data-tip="Add to Wishlist"><i class="fa fa-heart"></i></a></li>
-                      <li><a href="main/shopingCart?id=<?=$id ?>" data-tip="Add to Cart" class="cart"  data-id='<?= $id; ?>' ><i class="fa fa-shopping-cart "></i></a></li>
+                      <li><a href="main/shopingCart?id='<?=$id ;?>'" data-tip="Add to Cart" class="cart"  data-id='<?= $id; ?>' ><i class="fa fa-shopping-cart "></i></a></li>
                   </ul>
                  
               </div>
               <div class="product-content">
-                  <h3 class="title"><a href="#"> <?php  echo $row->product_name ?></a></h3>
+                  <h3 class="title"><a href="#"> <?php  echo $row->product_short_desc ?></a></h3>
                   <div class="price">
                    $ <?php  echo $row->product_price ?>
                   </div>
@@ -225,8 +254,15 @@ echo "</div>";
                     // add to cart button
          // product id for javascript access
          echo "<div class='product-id display-none'></div>";
+        if(array_key_exists($id, $_SESSION['cart'])){
+            // echo "<a href='main/displayShopingCartItems' class='btn btn-success w-100-pct'>";
+            //     echo "Update Cart";
             echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
-        
+
+            echo "</a>";
+        }else{
+            echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
+        }
     ?>
                 </div>
           </div>
@@ -249,6 +285,53 @@ echo "</div>";
 </div>
 
 <!-- Product slider End -->
+
+<!-- start Adds -->
+<section class="three-column-images mt-5">
+  <div class="container " >
+    <div class="row">
+      <div class="col-xs-6 col-sm-4">
+        <img src="app/assets/images/add.jpg"
+             class="img-thumbnail img-responsive">
+        <a href="#" target="_blank"></a>
+         
+      </div>
+   
+       <div class="col-xs-6 col-sm-4">
+        <img src="app/assets/images/add2.jpg"
+             class="img-thumbnail img-responsive">
+        <a href="#" target="_blank"></a>
+        
+      </div>
+   
+      <div class="col-xs-6 col-sm-4">
+        <img src="app/assets/images/add.jpg"
+             class="img-thumbnail img-responsive">
+        <a href="#" target="_blank"></a>
+        
+      </div>
+      </div> 
+     </div>  
+  </section>
+<!-- End Adds -->
+
+
+<!-- Banner Start -->
+
+<!--div class="Add-banner">
+  <div class="container">
+    <div class="row">
+      <div class="banner">
+        <img src="app/assets/images/banner.jpg" alt="Banner">
+
+    </div>
+    
+  </div>
+</div>
+</div-->
+<!-- Banner End -->
+
+ 
 
 
 <!--Satrt Featured Product-->
@@ -278,12 +361,12 @@ echo "</div>";
                   <ul class="social">
                       <li><a href="main/product_details?action=product_details&product_id=<?PHP echo $id?>" data-tip="Quick View"><i class="fa fa-eye"></i></a></li>
                       <li><a href="main/wishlist?id=<?=$id?>" data-tip="Add to Wishlist"><i class="fa fa-heart"></i></a></li>
-
+                      <li><a href="main/shopingCart?id='<?=$id ;?>'" data-tip="Add to Cart" class="cart"  data-id='<?= $id; ?>' ><i class="fa fa-shopping-cart "></i></a></li>
                   </ul>
                  
               </div>
               <div class="product-content">
-                  <h3 class="title"><a href="#"> <?php  echo $row->product_name ?></a></h3>
+                  <h3 class="title"><a href="#"> <?php  echo $row->product_short_desc ?></a></h3>
                   <div class="price">
                   $ <?php  echo $row->product_price ?>
 
@@ -292,16 +375,20 @@ echo "</div>";
                     // add to cart button
          // product id for javascript access
          echo "<div class='product-id display-none'></div>";
+        if(array_key_exists($id, $_SESSION['cart'])){
+           // echo "<a href='main/displayShopingCartItems' class='btn btn-success w-100-pct'>";
+           echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
+
+               // echo "Update Cart";
+            echo "</a>";
+        }else{
             echo "<a  class='add-to-cart' href='main/shopingCart?id={$id}' class='btn btn-primary w-100-pct'>أضف الى عربة التسوق</a>";
-        
+        }
     ?>
                 </div>
           </div>
       </div>
-      <?php //$i++;
-      //  } 
-    }
-    ?> 
+      <?php  } ?> 
 
   </div>
 </div>
@@ -317,18 +404,10 @@ echo "</div>";
 
 <!--End View Moer-->
 
-<!--Start View Moer-->
-<div class="overview text-center mt-5">
-         <div class="container">
-          
-           <button><a href="main/getcat"> عرض المزيد</a></button>
-          </div>
-       </div>
 
-<!--End View Moer-->
-
-
-
+<!-- start Adds -->
+<section class="three-column-images mt-5">
+  <div class="container " >
     <div class="row">
       <div class="col-xs-6 col-sm-4">
         <img src="http://localhost/Ecom-store-project/app/assets/images/add.jpg"
@@ -350,17 +429,38 @@ echo "</div>";
         <a href="#" target="_blank"></a>
         
       </div>
+      </div> 
+     </div>  
+  </section>
+<!-- End Adds -->
+
+
+
+
+      
+  <!-- <div class="container">
+	<div class="row">
+		<div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" width="100%" /></div>
+        <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+        <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+        <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+    	<div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+        <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+	    <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+        <div class="col-md-3 col-sm-4 col-xs-6"><img class="img-responsive" src="/Ecom-store-project/app/assets/images/barndsam.png" /></div>
+    </div>
+</div> -->
+   
+
+
 
 
 
 
 
     
- 
 <?PHP
 include "app/views/footer.php"; 
-?>  
-
-
-</body>
+?>   
+  </body>
 </html>
